Continued refactoring for Repositories
Same as previous commit
